Initial changes; need to work on physics some

+++ src/Maverick/Camera/Camera.hpp
@@ -17,8 +17,8 @@ class Camera{
 		~Camera();
 
 		void setReference(Object* reference);
-		void _draw(ObjectList* obj_lst, uint32_t delta, double camera_x_offset = 0, double camera_y_offset = 0);
-		void _draw(UIElementList* element_lst, uint32_t delta);
+		void _draw(ObjectList* obj_lst, uint64_t delta, double camera_x_offset = 0, double camera_y_offset = 0);
+		void _draw(UIElementList* element_lst, uint64_t delta);
 
 		SDL_Renderer* getRenderer();
 		double getX();
