+++ src/Maverick/Entity/UI/UIObjectFrame/UIObjectFrame.hpp
@@ -10,7 +10,7 @@ class UIObjectFrame : public UIElement{
         UIObjectFrame(const char* name, double view_x_offset, double view_y_offset, double view_width, double view_height, SDL_Window* window, int draw_layer, Object* reference = nullptr, char* border_pattern = "", uint8_t border_types = BORDER_TYPE::ALL_BORDERS, int border_buff = 0);
         ~UIObjectFrame();
 
-        virtual void draw(SDL_Renderer* renderer, uint32_t delta, int camera_x, int camera_y);
+        virtual void draw(SDL_Renderer* renderer, uint64_t delta, int camera_x, int camera_y);
 
         void setReference(Object* reference);
