+++ src/Maverick/Entity/UI/UIText/UIText.hpp
@@ -15,8 +15,8 @@ class UIText : public UIElement{
         UIText(const char* name, double view_x_offset, double view_y_offset, double view_width, double view_height, SDL_Window* window, int draw_layer, const char* font_path, const char* text = "", float scroll_speed = 0.0, unsigned int point = 12, ALIGNMENT x_alignment = ALIGNMENT::NORMAL, ALIGNMENT y_alignment = ALIGNMENT::NORMAL);
         virtual ~UIText();
 
-        void draw(SDL_Renderer* renderer, uint32_t delta, int camera_x, int camera_y);
-        void process(uint32_t delta);
+        void draw(SDL_Renderer* renderer, uint64_t delta, int camera_x, int camera_y);
+        void process(uint64_t delta);
 
         //Set the displayed text
         void setText(const char* text);
